Adding doc, new structure and data test

+++ backend/src/main/java/com/dasher/logisticsProfitSystem/adapter/rest/dto/request/ProfitRequest.java
@@ -2,10 +2,11 @@ package com.dasher.logisticsProfitSystem.adapter.rest.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProfitRequest(
     @NotNull UUID shipmentId,
-    @NotNull Integer income,
-    @NotNull Integer costs,
-    @NotNull Integer additionalCosts) {}
+    @NotNull BigDecimal income,
+    @NotNull BigDecimal costs,
+    @NotNull BigDecimal additionalCosts) {}
